test: refactor tests for GitTasks
